Refactor timing counting & display
- Hammer ISR should set start_time
- ISR for MPU6050 should set impact_time
- Elapsed time not yet set, but should be simple when the time comes.

However, not working yet:  for some reason, start time does not seem
to be set -- or at least, not displayed on screen or serial printing,
even when printed in main loop.

Even when I just touch the pad gently -- avoiding the case where the
hammer interrupt and the mpu interrupt are competing for CPU time -- I
don't see the start time changing.

I *do* see state changing to 2 when the hammer makes contact.  And the