Rework pulsedelay so that I can measure two times at once 🥳
This adds a new method, activate(), which is a thin wrapper around the
statemachine init() and active() calls.  Previously, both of these had
been in the get() method -- meaning that I was effectively only able
to take one measurement at a time, rather than having both collecting
values in the background until I was able to collect them in main().