Break out headers to a shared module
Since we're now actually printing out CSV, we should have the headers
printed out as well.

Originally, my intent was:

- Have the logger reset the micropython process when it started, by
  sending Ctrl-C/Ctrl-D to the serial port

- Have the micropython firmware print out the headers when it started

- Headers are now printed out by the micropython, they're read in by
  the logger, and there's no need to duplicate them across the two
  codebases

- Profit!

However, the logger doesn't appear to be reading the headers.  I'm not
sure what's going wrong, since the rest of the process is working.

To get around this:  I've broken out the DEFAULT_HEADERS variable to
the constants module in the micropython directory, and am importing it
from the logger.  At least that means there's only one place to update
this...